@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -28,15 +28,18 @@ dependencies:
 
   flutter_neumorphic: ^3.0.2
   footer: ^0.0.3
-  url_launcher: ^5.5.3
-  firebase: ^7.3.0
+  # url_launcher: ^5.5.3
+  # firebase: ^7.3.0
+  url_launcher: ^6.0.19
+  firebase: ^9.0.2
+  cupertino_icons: ^1.0.4
 
   flutterboilerplate:
     path: /Users/salvatore/Desktop/DevelopmentProjects/FlutterBoilerplate/flutter_boilerplate
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
+  # cupertino_icons: ^0.1.3
 
 dev_dependencies:
   flutter_test:
